fix all silly bugs

+++ app.js
@@ -4,7 +4,6 @@ const express = require('express');
 const Sequelize = require('sequelize');
 
 dotenv.load();
-const PORT = process.env.PORT || 3000;
 const db = {
     pass: process.env.POSTGRES_PASS,
     user: process.env.POSTGRES_USER,
@@ -18,20 +17,18 @@ const sequelize = new Sequelize(`postgres://${db.user}:${db.pass}@${db.host}:543
 const bodyParser = require('body-parser');
 const app = express();
 
-app.use(bodyParser({ urlencoded: true }));
+app.use(bodyParser.urlencoded({ extended: false }));
 
-const User = sequelize.define('user', {
-    username: Sequelize.STRING,
+const User = sequelize.define('Users', {
+    username: {
+        type: Sequelize.STRING,
+        unique: true,
+    },
     password: Sequelize.STRING,
 });
 
 //? Only need to run this once
-// User.sync().then(() => {
-//     User.create({
-//         username: 'jimmy',
-//         password: 'bananapancake',
-//     });
-// })
+User.sync();
 
 // send empty response on homepage
 app.get('/', (req, res) => {
@@ -45,40 +42,56 @@ app.get('/users', (req, res) => {
     })
 })
 
-// get single user by url param (i.e. '/user/lachlan')
+// get single user by url param (i.e. '/user/jimmy')
 app.get('/user/:username', (req, res) => {
     User.findOne({
         where: {
             username: req.params.username
         }
     }).then(users => {
-        res.json(users.rows);
+        res.json(users || {});
     })
 })
 
 // create new user via post data
-app.post('/create', (req, res) => {
+app.post('/user', (req, res) => {
+    if (!req.body.username) return res.json({ 'error': 'must include username' });
+    if (!req.body.password) return res.json({ 'error': 'must include password' });
     User.create({
         username: req.body.username,
         password: req.body.password,
+    })
+    .then(user => {
+        res.json(user);
+    })
+    .catch(err => {
+        res.json({
+            error: err.errors ? err.errors[0].message : `Error creating user ${req.body.username}`
+        })
     });
 });
 
 // update existing user via post data
-app.post('/update', (req, res) => {
+app.put('/user', (req, res) => {
+    if (!req.body.username) return res.json({ 'error': 'must include username' });
+    if (!req.body.password) return res.json({ 'error': 'must include password' });
     User.update({
         'password': req.body.password,
     }, {
-            where: {
-                username: req.body.username,
-            }
-        }).then(user => {
-            res.json({ 'success': true });
-        })
+        where: {
+            username: req.body.username,
+        }
+    }).then(user => {
+        res.json({ 'success': true });
+    })
+    .catch((err) => {
+        return res.json({ 'error': err || `failed to update user ${req.body.username}` })
+    })
 });
 
 // delete user via post data
-app.post('/delete', (req, res) => {
+app.delete('/user', (req, res) => {
+    if (!req.body.username) return res.json({'error':'must include username'});
     User.destroy({
         where: {
             username: req.body.username,
@@ -86,8 +99,13 @@ app.post('/delete', (req, res) => {
     }).then(user => {
         res.json({ 'success': true });
     })
+    .catch((err) => {
+        return res.json({'error': err || `failed to delete user ${req.body.username}`})
+    })
 });
 
+//set app to listen on PORT env var or default to 3000
+const PORT = process.env.PORT || 3000;
 app.listen(PORT, () => {
-    console.log('listening on port 3000');
+    console.log(`listening on port ${PORT}`);
 })
